Refactoring to move FilterBar in home package.
